Build - make misc updates for having precompiled tau assets in build

+++ app/linux-clean.sh
@@ -12,11 +12,13 @@ rm -rf external/build
 
 echo "Cleaning out BEAM distribution..."
 rm -rf server/beam/tau/_build
-rm -rf server/beam/tau/priv/static/assets
 rm -rf server/beam/tau/priv/*.{so,dylib,dll}
-rm -rf server/beam/tau/priv/static/cache_manifest.json
-find . -path './server/beam/tau/priv/static/*' -name '*.gz' -delete
-find . -path './server/beam/tau/priv/static/*' -name '*-????????????????????????????????.*' -delete
+if [ "$MIX_ENV" == dev ]; then
+    rm -rf server/beam/tau/priv/static/assets
+    rm -rf server/beam/tau/priv/static/cache_manifest.json
+    find . -path './server/beam/tau/priv/static/*' -name '*.gz' -delete
+    find . -path './server/beam/tau/priv/static/*' -name '*-????????????????????????????????.*' -delete
+fi
 
 echo "Cleaning completed"
